+++ CHANGELOG.md
@@ -14,6 +14,14 @@ Please add your changelog entry under this comment in the correct category (Secu
 ### Fixed
 * Fix deprecation warning for set-output in internal docker-scan job
 
+### Added
+* Register optional json log format @eiabea (#189)
+
+### Changes
+* upgrade to Go 1.20 @eiabea (#189)
+* dependency update
+  - Bump k8s.io/cloud-provider from 0.26.3 to 0.27.0 @dependabot (#189)
+
 ## [1.5.3] - 2023-04-04
 
 ### Fixed